api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic[email]==2.10.5
python-multipart==0.0.20
python-dotenv==1.0.1
pypdf==5.1.0
Pillow==11.1.0
pytest==8.3.4
